chore(kadena-cli): add option to config init in home directory

+++ packages/tools/kadena-cli/src/account/commands/accountDelete.ts
@@ -2,6 +2,7 @@ import { join } from 'path';
 import { NO_ACCOUNTS_FOUND_ERROR_MESSAGE } from '../../constants/account.js';
 import { ACCOUNT_DIR } from '../../constants/config.js';
 import { services } from '../../services/index.js';
+import { KadenaError } from '../../services/service-error.js';
 import type { CommandResult } from '../../utils/command.util.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { createCommand } from '../../utils/createCommand.js';
@@ -11,6 +12,9 @@ import { accountOptions } from '../accountOptions.js';
 import { ensureAccountAliasFilesExists } from '../utils/accountHelpers.js';
 
 async function deleteAccountDir(): Promise<CommandResult<null>> {
+  if (ACCOUNT_DIR === null) {
+    throw new KadenaError('no_kadena_directory');
+  }
   try {
     await services.filesystem.deleteDirectory(ACCOUNT_DIR);
     return {
@@ -28,6 +32,9 @@ async function deleteAccountDir(): Promise<CommandResult<null>> {
 async function removeAccount(
   accountAlias: string,
 ): Promise<CommandResult<null>> {
+  if (ACCOUNT_DIR === null) {
+    throw new KadenaError('no_kadena_directory');
+  }
   if (accountAlias === 'all') {
     return await deleteAccountDir();
   }
